requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.2.0
fake-useragent==1.1.3
flake8==7.0.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.4
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2023.3
requests==2.31.0
typing_extensions==4.10.0
urllib3==2.0.3
yarl==1.9.2
